requirements.txt
# WB Manager - зависимости
flask>=3.0.0
aiohttp>=3.9.0
aiosqlite>=0.19.0
qrcode[pil]>=7.4
Pillow>=10.0.0
requests>=2.31.0
edge-tts>=6.1.0
pydub>=0.25.1
psutil>=5.9.0
python-telegram-bot>=20.0
pystray>=0.19.5
playwright>=1.45.0
beautifulsoup4>=4.12.0
nest_asyncio>=1.5.6
